Move stuff to service layer
