+++ api/bonsai_api/routers/memberships.py
@@ -2,7 +2,7 @@
 
 import logging
 
-from bonsai_api.crud.memberships import (get_groups_by_sample_ids,
+from bonsai_api.services.membership_service import (get_groups_by_sample_ids,
                                          get_samples_by_group_ids)
 from bonsai_api.db import Database
 from bonsai_api.dependencies import get_current_active_user, get_database
@@ -60,5 +60,5 @@ async def get_group_membership(
         )
 
     if sample_ids:
-        return await get_groups_by_sample_ids(sample_ids, db=db)
-    return await get_samples_by_group_ids(group_ids, db=db)
+        return await get_groups_by_sample_ids(db, sample_ids=sample_ids)
+    return await get_samples_by_group_ids(db, group_ids=group_ids)
